+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/machines/electric/FoodFabricator.java
@@ -20,9 +20,12 @@ public abstract class FoodFabricator extends AContainer {
 		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.WHEAT)}, new ItemStack[] {SlimefunItems.WHEAT_ORGANIC_FOOD});
 		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.CARROT)}, new ItemStack[] {SlimefunItems.CARROT_ORGANIC_FOOD});
 		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.POTATO)}, new ItemStack[] {SlimefunItems.POTATO_ORGANIC_FOOD});
+		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.WHEAT_SEEDS)}, new ItemStack[] {SlimefunItems.SEEDS_ORGANIC_FOOD});
 		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.BEETROOT)}, new ItemStack[] {SlimefunItems.BEETROOT_ORGANIC_FOOD});
 		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.MELON)}, new ItemStack[] {SlimefunItems.MELON_ORGANIC_FOOD});
 		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.APPLE)}, new ItemStack[] {SlimefunItems.APPLE_ORGANIC_FOOD});
+		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.SWEET_BERRIES)}, new ItemStack[] {SlimefunItems.SWEET_BERRIES_ORGANIC_FOOD});
+		registerRecipe(12, new ItemStack[] {SlimefunItems.CAN, new ItemStack(Material.DRIED_KELP)}, new ItemStack[] {SlimefunItems.KELP_ORGANIC_FOOD});
 	}
 	
 	@Override