+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Added String to the Bio Reactor
 * Added a config setting to limit how many Nodes a Network can have
 * Added support for Furnaces and Brewing Stands to Cargo Networks
+* Added Organic Food/Fertilizer for Dried Kelp
 
 ### Changes
 * Removed Solar Array
@@ -50,6 +51,7 @@
 * Fixed #1379
 * Fixed #1212
 * Fixed #114
+* Fixed #1385
 
 ## Release Candidate 4 (06 Jan 2020)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#4